MM-40437 Move UserGuideDropdown component
